cybermorph_backend/requirement.txt
fastapi
uvicorn
lightgbm
numpy
sqlalchemy
pydantic
passlib[bcrypt]
pyjwt
python-multipart
